Properly implement all the glcolor functions
Note that GX ultimately only uses u8's, so we might as well truncate right after the glcolor call (though I think gcc maybe smart enough to save memory automatically with -o3... lets do this to be safe and simply things)

I programmed this based on the opengl documentation I read, which said:
A colour variable all 1's (i.e. 0xFF for u8, 0x7FFF for s16) is equivalent to a float/double of 1.0. Based on this information, I truncated u16/u32 and multiplied by 0xFF (255) for floats/doubles to get an equalent u8 value. Note that signed variables use the leftmost bit as a sign bit, so I shifted all positives left one, to remove the sign bit and to not lose precision before truncation. All negatives are set to 0 as well.

If any of this is wrong, please feel free to change this up.